creating list avatar characters
