+++ lib/app.dart
@@ -1,8 +1,7 @@
-import 'package:bloc_avatar_app/avatarFirstPart/screens/homeScreen/home.dart';
+import 'package:bloc_avatar_app/avatarFirstPart/screens/homeScreen/home_screen.dart';
 import 'package:flutter/material.dart';
 
 class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
